Litt usikker på ka æ har gjort, men sikkert nå lurt ;)

+++ project/src/main/java/bookingSystem/Room.java
@@ -50,7 +50,7 @@ public class Room {
 		}
 	}
 	
-	public double getPricePerNight() {
+	public int getPricePerNight() {
 		return pricePerNight;
 	}
